pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aegis-pay"
version = "0.3.3"
description = "Semantic Payment Guardrail for AI Agents"
authors = [{name="Aegis Team"}]
readme = "README.pypi.md"
license = {text="MIT"}
requires-python = ">=3.10"
dependencies = [
    "pydantic",
    "python-dotenv",
]

[project.optional-dependencies]
llm = ["openai", "tenacity>=9.1.4"]
stripe = ["stripe"]
langchain = ["langchain-core"]
mcp = ["mcp>=1.26.0"]
dashboard = ["streamlit>=1.55.0"]
dev = ["pytest", "pytest-asyncio"]
browser = ["playwright>=1.40.0"]
all = ["aegis-pay[llm,stripe,langchain,mcp,dashboard,browser]"]

[tool.hatch.build.targets.wheel]
packages = ["aegis"]

[dependency-groups]
dev = [
    "vulture>=2.15",
]
